+++ SRC/Properties/SGResources.Designer.cs
@@ -61,7 +61,7 @@ namespace Solti.Utils.Proxy.Properties {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Type embedding is supported only in C# compilations..
+        ///   Looks up a localized string similar to Type embedding is supported only in C# 7.0+ compilations..
         /// </summary>
         internal static string LNG_NOT_SUPPORTED {
             get {